+++ react/AutoCompletedFields.test.js
@@ -113,11 +113,12 @@ describe('AutoCompletedFields', () => {
         'geolocationAutoCompleted',
         undefined
       )
-      expect(onChangeAddressArgument.city).toHaveProperty('value', city)
       expect(onChangeAddressArgument.state).toHaveProperty(
         'postalCodeAutoCompleted',
         undefined
       )
+      expect(onChangeAddressArgument.city).toBeFalsy()
+      expect(onChangeAddressArgument.neighborhood).toBeFalsy()
     })
 
     it('should not display country information', () => {