+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `neighborhood` info to the address summary for the country `ARG`.
+
+### Changed
+
+- Now different address fields can have equal `types` property.
+- The precedence present on `types` property will be respected.
+
 ## [3.17.1] - 2021-07-06
 
 ### Changed